Перемещение PersonPislService
